Restored deterministic hashing
Realized why you had np.random.seed in the __init__ call of the CSVec.
In order to allow the module that imports CSVec to be unimpeded in
setting a random seed for the whole run, I restore the preexisting
random state after generating the hashes. Also I renamed CS_VEC to have
a more standard capitalization/underscore use